+++ docker/raspios/build.sh
@@ -38,10 +38,12 @@ fi
 base_host=$(hostname)
 
 set -e
-echo "[$base_host] Building raspios_rootfs.tar.gz..."
-docker run --network=host --hostname=docker --privileged -it --rm -w /raspios -v /dev:/dev -v $(pwd):/app debian:buster-slim sh /app/build.sh rootfs
+if ! test -f "raspios_rootfs.tar.gz"; then
+  echo "[$base_host] Building raspios_rootfs.tar.gz..."
+  docker run --network=host --hostname=docker --privileged -it --rm -w /raspios -v /dev:/dev -v $(pwd):/app debian:buster-slim sh /app/build.sh rootfs
+fi
 echo "[$base_host] Building Docker image from rootfs..."
-docker build -t raspios:lite - <<EOF
+docker build -t raspios:lite . -f- <<EOF
 FROM scratch
 ADD ./raspios_rootfs.tar.gz /
 CMD ["bash"]
